+++ src/pages/Account/AccountOrders/Order/OrderModal/index.module.scss
@@ -31,7 +31,6 @@
 
   .list_wrapper {
     padding: 2rem 3rem 0;
-
     width: 100%;
     display: flex;
     flex-direction: column;
@@ -88,7 +87,6 @@
 
         .image {
           max-width: 100%;
-          border: 0.1rem solid #cfcfcf;
           border-radius: 1rem;
           overflow-x: hidden;
           align-self: center;
