--- src/pages/Cart/CartContent/index.module.scss
@@ -1,156 +1,0 @@
-@import 'App.scss';
-
-.container {
-  padding: 10rem 0 4rem 0;
-}
-
-.title {
-  margin-bottom: 1rem;
-}
-
-.checkout_wrapper {
-  border: 0.1rem solid #cfcfcf;
-  border-radius: 1rem;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 2rem 2rem;
-  margin-bottom: 5rem;
-}
-
-.total {
-  @media (min-width: 768px) {
-    font-size: 2rem;
-  }
-}
-
-.checkout_button {
-  border: 0.1rem solid #000;
-  border-radius: 1rem;
-  text-transform: uppercase;
-  background: #000;
-  color: #fff;
-  padding: 1rem 4rem;
-  font-size: 1.6rem;
-  letter-spacing: $ls-button;
-
-  -moz-transition: all 0.1s ease-in;
-  -o-transition: all 0.1s ease-in;
-  -webkit-transition: all 0.1s ease-in;
-  transition: all 0.1s ease-in;
-
-  &:hover {
-    cursor: pointer;
-  }
-
-  &:active {
-    transform: scale(0.9);
-  }
-}
-
-.content_wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 3.5rem;
-
-  @media (min-width: 768px) {
-    display: grid;
-    grid-template-columns: 2fr 1fr;
-    gap: 4rem;
-  }
-}
-
-.list_wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 3rem;
-  padding-bottom: 4rem;
-  border-bottom: 0.2rem solid #cfcfcf;
-
-  @media (min-width: 768px) {
-    padding-right: 4rem;
-    border-right: 0.2rem solid #cfcfcf;
-    padding-bottom: 0;
-    border-bottom: none;
-  }
-}
-
-.support {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.support_title {
-  text-transform: uppercase;
-}
-
-.support_input {
-  border: solid 0.2rem #9e9e9e;
-  border-radius: 1rem;
-  padding: 1rem 2rem;
-  width: 100%;
-  font-size: 1.6rem;
-}
-
-.support_button {
-  padding: 1rem 2rem;
-  font-size: 1.6rem;
-  border: solid 0.2rem #000;
-  border-radius: 1rem;
-  width: 100%;
-  background: #000;
-  color: #fff;
-  text-transform: uppercase;
-  letter-spacing: $ls-button;
-
-  -moz-transition: all 0.1s ease-in;
-  -o-transition: all 0.1s ease-in;
-  -webkit-transition: all 0.1s ease-in;
-  transition: all 0.1s ease-in;
-
-  &:active {
-    transform: scale(0.9);
-  }
-}
-
-.no_products_wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 2rem;
-}
-
-.no_products {
-  text-align: center;
-}
-
-.products_button {
-  padding: 1rem 2rem;
-  font-size: 1.6rem;
-  border: solid 0.2rem #000;
-  border-radius: 1rem;
-  width: 100%;
-  background: #000;
-  color: #fff;
-  text-transform: uppercase;
-  letter-spacing: $ls-button;
-  color: white;
-
-  -moz-transition: all 0.1s ease-in;
-  -o-transition: all 0.1s ease-in;
-  -webkit-transition: all 0.1s ease-in;
-  transition: all 0.1s ease-in;
-
-  &:active {
-    transform: scale(0.9);
-  }
-
-  &:hover {
-    color: #fff;
-  }
-
-  @media (min-width: 768px) {
-    width: 50%;
-  }
-}
