+++ src/pages/Cart/index.module.scss
@@ -1,0 +1,222 @@
+@import 'App.scss';
+
+.container {
+  padding: 10rem 0 4rem 0;
+
+  .title {
+    margin-bottom: 1rem;
+    text-transform: uppercase;
+    font-size: 2rem;
+  }
+
+  .no_products_wrapper {
+    margin-top: 5rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 2rem;
+
+    .no_products {
+      text-align: center;
+      text-transform: uppercase;
+      font-size: 1.2rem;
+
+      @media (min-width: 360px) {
+        font-size: 1.3rem;
+      }
+
+      @media (min-width: 480px) {
+        font-size: 1.6rem;
+      }
+    }
+
+    .products_button {
+      padding: 1rem 2rem;
+      font-size: 1.2rem;
+      border: solid 0.2rem #000;
+      border-radius: 1rem;
+      width: 100%;
+      background: #000;
+      color: #fff;
+      text-transform: uppercase;
+      letter-spacing: $ls-button;
+      color: white;
+
+      -moz-transition: all 0.1s ease-in;
+      -o-transition: all 0.1s ease-in;
+      -webkit-transition: all 0.1s ease-in;
+      transition: all 0.1s ease-in;
+
+      &:active {
+        transform: scale(0.9);
+      }
+
+      &:hover {
+        color: #fff;
+      }
+
+      @media (min-width: 360px) {
+        font-size: 1.3rem;
+      }
+
+      @media (min-width: 480px) {
+        font-size: 1.6rem;
+      }
+
+      @media (min-width: 768px) {
+        width: 50%;
+      }
+    }
+  }
+
+  .checkout_wrapper {
+    border: 0.1rem solid #cfcfcf;
+    border-radius: 1rem;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 2rem 2rem;
+    margin-bottom: 5rem;
+
+    .total {
+      font-size: 1.2rem;
+      margin-right: 2rem;
+
+      span {
+        font-size: 1.4rem;
+      }
+    }
+
+    .total,
+    .total span {
+      text-transform: uppercase;
+
+      @media (min-width: 360px) {
+        font-size: 1.3rem;
+      }
+
+      @media (min-width: 480px) {
+        font-size: 1.6rem;
+      }
+    }
+
+    .checkout_button {
+      border: 0.1rem solid $border-c-button-checkout;
+      border-radius: 1rem;
+      text-transform: uppercase;
+      background: $back-c-button-checkout;
+      color: $c-secondary;
+      padding: 1rem 2rem;
+      font-size: 1.2rem;
+      letter-spacing: $ls-button;
+
+      -moz-transition: all 0.1s ease-in;
+      -o-transition: all 0.1s ease-in;
+      -webkit-transition: all 0.1s ease-in;
+      transition: all 0.1s ease-in;
+
+
+      @media (min-width: 480px) {
+        font-size: 1.6rem;
+        padding: 1rem 4rem;
+      }
+
+      &:hover {
+        border: 0.1rem solid $border-c-button-checkout-hover;
+        background: $back-c-button-checkout-hover;
+        cursor: pointer;
+      }
+
+      &:active {
+        transform: scale(0.9);
+      }
+    }
+  }
+
+  .content_wrapper {
+    display: flex;
+    flex-direction: column;
+    gap: 3.5rem;
+
+    @media (min-width: 768px) {
+      display: grid;
+      grid-template-columns: 2fr 1fr;
+      gap: 4rem;
+    }
+
+    .list_wrapper {
+      display: flex;
+      flex-direction: column;
+      gap: 3rem;
+      padding-bottom: 4rem;
+      border-bottom: 0.2rem solid #cfcfcf;
+
+      @media (min-width: 768px) {
+        padding-right: 4rem;
+        border-right: 0.2rem solid #cfcfcf;
+        padding-bottom: 0;
+        border-bottom: none;
+      }
+    }
+
+    .support {
+      display: flex;
+      flex-direction: column;
+      gap: 1rem;
+
+      .support_title {
+        text-transform: uppercase;
+
+        @media (min-width: 768px) {
+          font-size: 1.4rem;
+        }
+
+        @media (min-width: 900px) {
+          font-size: 1.6rem;
+        }
+      }
+
+      .support_input {
+        border: solid 0.2rem #9e9e9e;
+        border-radius: 1rem;
+        padding: 1rem 2rem;
+        width: 100%;
+        font-size: 1.6rem;
+      }
+
+      .support_button {
+        padding: 1rem 2rem;
+        font-size: 1.6rem;
+        border: solid 0.2rem $back-c-button-primary;
+        border-radius: 1rem;
+        width: 100%;
+        background: $back-c-button-primary;
+        color: #fff;
+        text-transform: uppercase;
+        letter-spacing: $ls-button;
+
+        -moz-transition: all 0.1s ease-in;
+        -o-transition: all 0.1s ease-in;
+        -webkit-transition: all 0.1s ease-in;
+        transition: all 0.1s ease-in;
+
+        @media (min-width: 768px) {
+          font-size: 1.4rem;
+        }
+
+        @media (min-width: 900px) {
+          font-size: 1.6rem;
+        }
+
+        &:active {
+          transform: scale(0.9);
+        }
+
+        &:hover {
+          background: $back-c-button-hover;
+          border: solid 0.2rem $back-c-button-hover;
+        }
+      }
+    }
+  }
+}
